Fixed MSVC 2010 compiler errors.

+++ src/dbi/driver/mysql/my_ddi.c
@@ -71,7 +71,9 @@ static BOOL __collect_column_definitions(CwtDDI *ddi, CwtDDIColumn *col, CwtStri
 		} else if( CWT_TYPE_IS_INTEGER(col->type) ) {
 			typestr = __stringifyIntType(col->opts.int_opts.min, col->opts.int_opts.max);
 		} else if(CWT_TYPE_IS_FLOAT(col->type)) {
-			typestr = __stringifyFloatType(col->type, col->opts.float_opts.prec, col->opts.float_opts.scale);
+			typestr = __stringifyFloatType(col->type
+					, (UINT)col->opts.float_opts.prec
+					, (UINT)col->opts.float_opts.scale);
 		} else if( CwtType_TEXT == col->type ) {
 			typestr = __stringifyTextType(col->opts.text_opts.maxlen);
 		} else if( CwtType_BLOB == col->type ) {
